ass_render: add ass_set_event_types, allowing split dialogue/TS rendering
At any given timestamp, we consider each event to be either dialogue or TS.
All dialogue is always layered above all TS, so it's safe to blend all TS
prior to any dialogue. This can be used to render TS at the video's resolution
prior to scaling for display (improving render consistency and, if upscaling,
performance), then render dialogue at the output resolution (maintaining
crisp sharp output for ordinary subtitle text).

The implementation is fairly simple: a renderer can be configured to render
either only dialogue, only TS, or both. Events are split between the two passes
based on their layer. The minimum layer that's considered "dialogue" is
the lowest-numbered layer on which an event exists that passes a heuristic,
which considers an event "probably dialogue" if it would trigger collision
handling. This means that all events that require collision handling are always
handled in the dialogue renderer, so we don't need to worry about collisions
split across multiple renderers, or colliding events moving between renderers
on different frames.

The main caveat this introduces is that it's possible for an event to be
considered dialogue on one frame and typesetting on another, or vice versa.
This can happen if an event that passes the dialogue heuristic is layered
beneath other events that don't, if those other events are onscreen for longer.
This generally should have fairly minor impact as long as our render scaling
is reasonably correct.